+++ apps/carp_mobile_sensing_app/lib/src/sensing/local_study_protocol_mananger.dart
@@ -9,17 +9,36 @@ part of mobile_sensing_app;
 
 /// This is a simple local [StudyProtocolManager].
 ///
-/// This class shows how to configure a [StudyProtocol] with [Tigger]s,
+/// This class shows how to configure a [SmartphoneStudyProtocol] with [Tigger]s,
 /// [TaskDescriptor]s and [Measure]s.
 class LocalStudyProtocolManager implements StudyProtocolManager {
   Future initialize() async {}
 
   /// Create a new CAMS study protocol.
-  Future<StudyProtocol> getStudyProtocol(String studyId) async {
-    StudyProtocol protocol = StudyProtocol(
+  Future<SmartphoneStudyProtocol> getStudyProtocol(String studyId) async {
+    SmartphoneStudyProtocol protocol = SmartphoneStudyProtocol(
       ownerId: 'alex@uni.dk',
-      name: 'CAMS App - Sensing Coverage Study',
-      description: 'This is a protocol for testing the coverage of sampling.',
+      name: 'CAMS App Study #1',
+    );
+
+    protocol.protocolDescription = StudyDescription(
+        title: 'CAMS App - Sensing Coverage Study',
+        description:
+            'The default study testing coverage of most measures. Used in the coverage tests.',
+        purpose: 'To test sensing coverage',
+        responsible: StudyReponsible(
+          id: 'jakba',
+          title: 'professor',
+          address: 'Ørsteds Plads',
+          affiliation: 'Technical University of Denmark',
+          email: 'jakba@dtu.dk',
+          name: 'Jakob E. Bardram',
+        ));
+
+    // add CARP as the data endpoint - w/o authentication info - we expect to be authenticated
+    protocol.dataEndPoint = CarpDataEndPoint(
+      uploadMethod: CarpUploadMethod.DATA_POINT,
+      name: 'CARP Server',
     );
 
     // Define which devices are used for data collection.