+++ utilities/carp_study_generator/test/carp_study_generator_test.dart
@@ -35,7 +35,6 @@ void main() {
   group("commands", () {
     test('help', () async => HelpCommand().execute());
     test('dryrun', () async => DryRunCommand().execute());
-    test('description', () async => StudyDescriptionCommand().execute());
     test('consent', () async => ConsentCommand().execute());
     test('protocol', () async => CreateStudyProtocolCommand().execute());
   });