+++ backends/carp_webservices/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.31.0
+* upgrade to `carp_mobile_sensing` v. 0.31.0
+
 ## 0.30.1
 * fix of bug in `DocumentReference.updateData()` method.
 * Dialog windows for authentication and selection of study invitation can now be made "modal" (i.e., not closable by the user).